add authentication & authorization for  aws cognito

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StorePostRequest;
 use App\Post;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
@@ -29,6 +31,9 @@ class PostController extends Controller
 
     public function update(Request $request, Post $post): JsonResponse
     {
+        // dump(Auth::user());
+        Gate::authorize('admin-req', $request);
+
         $post->fill($request->all());
         $post->save();
 
@@ -37,6 +42,9 @@ class PostController extends Controller
 
     public function destroy(Post $post): JsonResponse
     {
+        // dump(Auth::user());
+        Gate::authorize('manager');
+
         $post->delete();
 
         return response()->json($post);
